Remove all lingering references to EJS and reformat template import/export structure
